api_server/requirements.txt
fastapi
uvicorn
httpx
trio
pytest
pymongo
pydantic
python-multipart
Pillow